+++ CHANGELOG.md
@@ -1,29 +1,117 @@
-# [2.2.0](https://github.com/PPL-B3/avento-origin-fe/compare/v2.1.0...v2.2.0) (2025-04-16)
+# [2.15.0](https://github.com/PPL-B3/avento-origin-fe/compare/v2.14.0...v2.15.0) (2025-04-28)
 
 
 ### Bug Fixes
 
-* [BLUE] adjust code ([09151e8](https://github.com/PPL-B3/avento-origin-fe/commit/09151e8dcbd79e07bdca98060a4c3e48e88c2aa7))
-* [BLUE] fix typo ([762c2fb](https://github.com/PPL-B3/avento-origin-fe/commit/762c2fb1889ebf6c23aed3e026fcb39e1dea6ff2))
-* authbutton object ([dc0d9c9](https://github.com/PPL-B3/avento-origin-fe/commit/dc0d9c979c6f687e8128f2049b42e4d544e0dab6))
-* code duplication ([344a3cd](https://github.com/PPL-B3/avento-origin-fe/commit/344a3cd05f7e0ab3654ddd45d6aa068dd8ba6dc2))
-* Coderabbit follow up issue ([083d467](https://github.com/PPL-B3/avento-origin-fe/commit/083d4673a7c6958acd73bbf9910d64728439f109))
-* coderabbit issue ([3ff1be3](https://github.com/PPL-B3/avento-origin-fe/commit/3ff1be367d06f9894fd5dbe371b7df6a0fe3d227))
-* lint and sonarqube duplication ([5458cde](https://github.com/PPL-B3/avento-origin-fe/commit/5458cdeac37a233b2e0b1a443ad69f24da273837))
-* sonarqube issues ([8c34571](https://github.com/PPL-B3/avento-origin-fe/commit/8c34571d0944cb91302ef1ea70e08ef8d107329a))
-* tests ([f331a09](https://github.com/PPL-B3/avento-origin-fe/commit/f331a09523ccb9ed64987d6e91e5fcef6934f1d2))
+* add encrypt email on transfer history ([c04c1d5](https://github.com/PPL-B3/avento-origin-fe/commit/c04c1d50f7ab6f9a794a57a0fde55fc3200072df))
+* lint error unused ([3ebba67](https://github.com/PPL-B3/avento-origin-fe/commit/3ebba676b3b9352cb8bf8e0a781becd364dc2cbb))
+* lint unused ([8ed951a](https://github.com/PPL-B3/avento-origin-fe/commit/8ed951ad7070ad7c525f36308663081d88d3226a))
+* pagination ([c1fa972](https://github.com/PPL-B3/avento-origin-fe/commit/c1fa9720071c6054c8bf0019e54023599533d030))
+* remove deprecated ownerName ([29ee5f9](https://github.com/PPL-B3/avento-origin-fe/commit/29ee5f9ee4592a435be31425bbb10790aa333cd3))
+* reverse the order of transfer history ([ea07743](https://github.com/PPL-B3/avento-origin-fe/commit/ea0774305f6484efc99f8efcab3acdb54d9a732d))
+* search ([8c4690d](https://github.com/PPL-B3/avento-origin-fe/commit/8c4690d3d082e07203e8f20c384e3bc750a2ee8a))
+* wrap useSearchParams ([e7e8339](https://github.com/PPL-B3/avento-origin-fe/commit/e7e8339ec3a7aca4db6f8a03ce86326cb1821d75))
 
 
 ### Features
 
-* [BLUE] add use form ([8239092](https://github.com/PPL-B3/avento-origin-fe/commit/8239092d1827e6f2db7ca11f577204434d1bbd32))
-* [BLUE] adjust code ([b271acb](https://github.com/PPL-B3/avento-origin-fe/commit/b271acbfff1975fb3770aca7133b83afd8ae7310))
-* [BLUE] adjust components ([5c8856a](https://github.com/PPL-B3/avento-origin-fe/commit/5c8856a498af76b5d224c8553d7740c7130693d1))
-* [BLUE] fix code for the tests ([85fc5a3](https://github.com/PPL-B3/avento-origin-fe/commit/85fc5a3fa632aca241fc2f12c2fa0d1313e9a382))
-* [BLUE] move the validations ([a16fe56](https://github.com/PPL-B3/avento-origin-fe/commit/a16fe56f988134b05600f33e93835fa89a696b3e))
-* [RED] add tests ([b246cca](https://github.com/PPL-B3/avento-origin-fe/commit/b246ccada1924b219bb05f09a7847e5c91d884a0))
-* [RED] add tests for AuditLogModule ([90cf095](https://github.com/PPL-B3/avento-origin-fe/commit/90cf095426a30553bf75696f9270356620c1a072))
-* [RED] add transfer request tests ([3f7d9dd](https://github.com/PPL-B3/avento-origin-fe/commit/3f7d9dd66715f7d75802f72bd869448f116805dd))
+* add pagination logic ([cc4e89a](https://github.com/PPL-B3/avento-origin-fe/commit/cc4e89ac345c2bd37215c642d92762e2f704dc6f))
+* integrate filter and pagintation ([8d1d501](https://github.com/PPL-B3/avento-origin-fe/commit/8d1d501001385cd8d598b81bd2854c28d98da790))
+* integrate revert doc ([07049fa](https://github.com/PPL-B3/avento-origin-fe/commit/07049fac714d31654c5a5d31d15817cd83c652bd))
+
+# [2.14.0](https://github.com/PPL-B3/avento-origin-fe/compare/v2.13.0...v2.14.0) (2025-04-28)
+
+### Features
+
+- add hook for revert doc ([7d3ac7c](https://github.com/PPL-B3/avento-origin-fe/commit/7d3ac7c7172400b3d3d043306440e25052a3da82))
+- add temporary role ([55b304d](https://github.com/PPL-B3/avento-origin-fe/commit/55b304d430fb10e30aee3b8d0f1ff102cf720fc7))
+- integrate role access to BE ([66f5860](https://github.com/PPL-B3/avento-origin-fe/commit/66f586091c25d3366701acb152c4d45c5f87ca5a))
+
+# [2.13.0](https://github.com/PPL-B3/avento-origin-fe/compare/v2.12.0...v2.13.0) (2025-04-28)
+
+### Features
+
+- [GREEN] implement test ([e6c66c4](https://github.com/PPL-B3/avento-origin-fe/commit/e6c66c4b111533411295365bea02f4ecc5f2f454))
+- **metadata:** [RED] add new tests for document detail page ([aac55ff](https://github.com/PPL-B3/avento-origin-fe/commit/aac55ff96901877eadb968a0d4dd81260017f97e))
+- **metadata:** [REFACTOR] conditional rendering document detail page via otp verification ([629bc0e](https://github.com/PPL-B3/avento-origin-fe/commit/629bc0e67b42f64f34c407ef0313d63cbc4d32d8))
+- **metadata:** [REFACTOR] implement conditional rendering document detail page ([1b2a939](https://github.com/PPL-B3/avento-origin-fe/commit/1b2a939c1488e56188a00967a3095af0bc39e9bb))
+
+# [2.12.0](https://github.com/PPL-B3/avento-origin-fe/compare/v2.11.0...v2.12.0) (2025-04-28)
+
+### Features
+
+- add endpoint document access ([5b94206](https://github.com/PPL-B3/avento-origin-fe/commit/5b942069178a05904f94b959a48fc18a276f557b))
+- add hook for OTP verification to access document API integration ([cddd917](https://github.com/PPL-B3/avento-origin-fe/commit/cddd917c5d510ffcb82d8bcd030fda53cd39fed2))
+- add hook for request document access API integration ([f41cc14](https://github.com/PPL-B3/avento-origin-fe/commit/f41cc1409eb3f52be60cbc138755959b90fbac09))
+
+# [2.11.0](https://github.com/PPL-B3/avento-origin-fe/compare/v2.10.0...v2.11.0) (2025-04-28)
+
+### Bug Fixes
+
+- husky ([562aa61](https://github.com/PPL-B3/avento-origin-fe/commit/562aa6128468b379f503bb83cd48b35de1862dee))
+- lint error ([de6256e](https://github.com/PPL-B3/avento-origin-fe/commit/de6256e5180eb52274bff6e32be9069981970997))
+- lint error import ([9631db8](https://github.com/PPL-B3/avento-origin-fe/commit/9631db8f482fa35f6653afd77f3d8d4f60ae63fc))
+
+### Features
+
+- add admin log document type ([b85d0b3](https://github.com/PPL-B3/avento-origin-fe/commit/b85d0b395fb3d3ee24f0db2599b1d16d60ddbb5b))
+- add history type ([558c281](https://github.com/PPL-B3/avento-origin-fe/commit/558c281c583fb713e72d0af25eb1c167945dbc98))
+- add hook for fetch admin doc detail ([e92299b](https://github.com/PPL-B3/avento-origin-fe/commit/e92299be8fe05d88082b1f88d03e962de9f1d366))
+
+# [2.10.0](https://github.com/PPL-B3/avento-origin-fe/compare/v2.9.0...v2.10.0) (2025-04-28)
+
+### Bug Fixes
+
+- disable transfer and view document ([4852e56](https://github.com/PPL-B3/avento-origin-fe/commit/4852e56c95e819969244e93c918cfc89a52a5cb7))
+
+### Features
+
+- add dummy data ([7ca7987](https://github.com/PPL-B3/avento-origin-fe/commit/7ca798705cc78eacf59f2c04b3a5814964cb9d98))
+- add error handling ([7bdcd67](https://github.com/PPL-B3/avento-origin-fe/commit/7bdcd67a085cd1295887f66ceebffcae78c64d11))
+- add toast for revert ([5a7cef2](https://github.com/PPL-B3/avento-origin-fe/commit/5a7cef2af00e1d2360a0cf8cb0aa66a7b6f55d5f))
+
+# [2.9.0](https://github.com/PPL-B3/avento-origin-fe/compare/v2.8.0...v2.9.0) (2025-04-28)
+
+### Features
+
+- add detail audit log ([325ca11](https://github.com/PPL-B3/avento-origin-fe/commit/325ca1103cb57d0dba38375d815bf461100859e1))
+- add document detail ([df18f96](https://github.com/PPL-B3/avento-origin-fe/commit/df18f96fd29395fda1eaed292b6962e260836902))
+
+# [2.8.0](https://github.com/PPL-B3/avento-origin-fe/compare/v2.7.0...v2.8.0) (2025-04-28)
+
+### Bug Fixes
+
+- export ([8780150](https://github.com/PPL-B3/avento-origin-fe/commit/87801506c4896839141bb940326e98cfe20c9388))
+- folder structure ([b44c067](https://github.com/PPL-B3/avento-origin-fe/commit/b44c067683ad90cd8dea9e767c748c0414255b6d))
+
+### Features
+
+- iniitialize page ([fa0356a](https://github.com/PPL-B3/avento-origin-fe/commit/fa0356a8f15d409caaab909ebf85c2b317a33623))
+
+# [2.2.0](https://github.com/PPL-B3/avento-origin-fe/compare/v2.1.0...v2.2.0) (2025-04-16)
+
+### Bug Fixes
+
+- [BLUE] adjust code ([09151e8](https://github.com/PPL-B3/avento-origin-fe/commit/09151e8dcbd79e07bdca98060a4c3e48e88c2aa7))
+- [BLUE] fix typo ([762c2fb](https://github.com/PPL-B3/avento-origin-fe/commit/762c2fb1889ebf6c23aed3e026fcb39e1dea6ff2))
+- authbutton object ([dc0d9c9](https://github.com/PPL-B3/avento-origin-fe/commit/dc0d9c979c6f687e8128f2049b42e4d544e0dab6))
+- code duplication ([344a3cd](https://github.com/PPL-B3/avento-origin-fe/commit/344a3cd05f7e0ab3654ddd45d6aa068dd8ba6dc2))
+- Coderabbit follow up issue ([083d467](https://github.com/PPL-B3/avento-origin-fe/commit/083d4673a7c6958acd73bbf9910d64728439f109))
+- coderabbit issue ([3ff1be3](https://github.com/PPL-B3/avento-origin-fe/commit/3ff1be367d06f9894fd5dbe371b7df6a0fe3d227))
+- lint and sonarqube duplication ([5458cde](https://github.com/PPL-B3/avento-origin-fe/commit/5458cdeac37a233b2e0b1a443ad69f24da273837))
+- sonarqube issues ([8c34571](https://github.com/PPL-B3/avento-origin-fe/commit/8c34571d0944cb91302ef1ea70e08ef8d107329a))
+- tests ([f331a09](https://github.com/PPL-B3/avento-origin-fe/commit/f331a09523ccb9ed64987d6e91e5fcef6934f1d2))
+
+### Features
+
+- [BLUE] add use form ([8239092](https://github.com/PPL-B3/avento-origin-fe/commit/8239092d1827e6f2db7ca11f577204434d1bbd32))
+- [BLUE] adjust code ([b271acb](https://github.com/PPL-B3/avento-origin-fe/commit/b271acbfff1975fb3770aca7133b83afd8ae7310))
+- [BLUE] adjust components ([5c8856a](https://github.com/PPL-B3/avento-origin-fe/commit/5c8856a498af76b5d224c8553d7740c7130693d1))
+- [BLUE] fix code for the tests ([85fc5a3](https://github.com/PPL-B3/avento-origin-fe/commit/85fc5a3fa632aca241fc2f12c2fa0d1313e9a382))
+- [BLUE] move the validations ([a16fe56](https://github.com/PPL-B3/avento-origin-fe/commit/a16fe56f988134b05600f33e93835fa89a696b3e))
+- [RED] add tests ([b246cca](https://github.com/PPL-B3/avento-origin-fe/commit/b246ccada1924b219bb05f09a7847e5c91d884a0))
+- [RED] add tests for AuditLogModule ([90cf095](https://github.com/PPL-B3/avento-origin-fe/commit/90cf095426a30553bf75696f9270356620c1a072))
+- [RED] add transfer request tests ([3f7d9dd](https://github.com/PPL-B3/avento-origin-fe/commit/3f7d9dd66715f7d75802f72bd869448f116805dd))
 
 # [2.1.0](https://github.com/PPL-B3/avento-origin-fe/compare/v2.0.1...v2.1.0) (2025-04-10)